Selectively count stack slots when reporting StkSizeRecord for Stack Map
LLVM might preallocate stack space for parameter passing, so when a function
is called, no push is needed, params go the preallocated stack slots.

For the purposes of safepoint generation in LLV8 we need to know the number
of stack slots excluding the parameter-slots mentioned above (as well as
alignment bytes).